Add flag to object creation count - Only count after intrinsics are initialized
