+++ src/values/ObjectValue.js
@@ -69,7 +69,7 @@ export default class ObjectValue extends ConcreteValue {
     this.properties = new Map();
     this.symbols = new Map();
     this.refuseSerialization = refuseSerialization;
-    realm.interpreterStatistics.objects++;
+    if (realm.intrinsicsInitialized) realm.interpreterStatistics.objects++;
   }
 
   static trackedPropertyNames = [
